fix(ci): keep scope discovery import-light
